NLP_chatbot/requirement.txt
mysql-connector-python==8.0.33
langchain-community==0.0.38
langchain==0.1.20
requests==2.31.0
